Move get_current_shot to base qis so utilities can access it without binding specifically to helios_qis. Correct rpath of helios_qis for installs so that base_qis can be found.
