+++ selene-ext/interfaces/helios_qis/c/include/helios_qis/helios_ops.h
@@ -22,7 +22,6 @@ EXPORT bool ___read_future_bool(uint64_t r);
 EXPORT void ___barrier(uint64_t* qubits, uint64_t qubits_len);
 EXPORT void ___sleep(uint64_t* qubits, uint64_t qubits_len, uint64_t sleep_time);
 EXPORT uint64_t ___read_future_uint(uint64_t r);
-EXPORT uint64_t get_current_shot();
 EXPORT void set_tc(uint64_t time_cursor);
 EXPORT uint64_t get_tc();
 EXPORT void setup(uint64_t time_cursor);
